requirements.txt
appdirs==1.4.4
asgiref==3.2.8
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
dj-stripe==2.3.0
dj_database_url==0.5.0
Django==3.0.7
django-cors-headers==3.4.0
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
flake8==3.8.3
Flask==1.1.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jsonfield==3.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
ortools==7.6.7691
pathspec==0.8.0
protobuf==3.12.2
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pytz==2020.1
regex==2020.6.8
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
stripe==2.48.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
